gui2/tslider: Improved color scheme for slider grooves
This makes it so the "selected" segment of the groove (towards the
slider origin) is drawn with a brighter color than the "empty" segment
(opposite from the origin). Since this obviously wouldn't work with
RGB #FFFFFF, and the color is too distracting anyway (it was pretty much
the only widget using the general *text* color), we now use the standard
UI border colors to match buttons, checkboxes and so on. This should
hopefully help with harmonizing Wesnoth's UI color scheme across the
board.

For this purpose we also introduce a macro for a darker/shaded disabled
text color, used for the unselected segment of disabled sliders.


Former-commit-id: 293bbcd1c67f4f620b34ef9d4c616e14de6e3fad
Former-commit-id: d036013c3f802ec071006471801c00b8e35cd7fc
Former-commit-id: 18ead64a00d7131287706e7b087b3742d7c96e7f
Former-commit-id: 284fcb4c9108b69c199cafa25335e36595c76ccf [formerly acb08e60c9b4479d3c739485ecf426c378567cf3]
Former-commit-id: 2ad0e08e2ac4904ee8c066e34eb2ffe78e1e36c1