MP Staging: fixed crash involving gold/income sliders
For some reason, changing the slider and canceling the dialog caused a crash. Changing the value
while in a networked game also caused a crash. If I remember correctly, this did not occur when the
code used a refence to the side_engine directly, so I'm not sure why this crashed.


Former-commit-id: e3f237357c20e6a4cd92773aaf47285019e1c46a
Former-commit-id: a16fda219d566b5010400b499aadb53d9987b0f3
Former-commit-id: 87e0103b6fcbcb9610dc21994810f49d1cb67a58
Former-commit-id: 13c010c4dbcedd067fabc7e04cb9aab461dce631 [formerly 6a814d6b3fec6963aaeaa27b91bc55a58014f03d]
Former-commit-id: d099ba21f0a0ed300720adc67a5e6d05ef414909