gui2/canvas: Allow overriding color alpha in [line] with a formula
The color attribute specifically doesn't support formulas. Adding
support for formulas to it would present us with the issue that commas
are already part of the formula language, so we'd need to update all
existing instances accordingly. So, instead, there's this short hack
used for a very particular piece of functionality coming in the next
commit.


Former-commit-id: f8ba8c03536c27eb3cf50cfee145d77be61b0f7c
Former-commit-id: cdd0658970a9a3bd9d09dccb09a2336ef5f3b741
Former-commit-id: 840217d217415ac5bca88fbdb27dfa59d4ba3014
Former-commit-id: 4ab986be63556249f5d04127cb09e0ceef1a61d9 [formerly 300b4a3b00994baa73a2a79f61001d08fc366024]
Former-commit-id: a7dc982f8f2e0bb871985ee39192de7eaa112def