gui2/ttext_: Add a blinking cursor to textbox widgets
This is used by all widgets that inherit from the textbox abstract
class, ttext_. The cursor is updated every 750 ms by toggling the
cursor alpha every time. When the cursor position is set by a method
(e.g. in reaction to a keyboard event), the cursor is reset to full
alpha, and the blink timer restarted.


Former-commit-id: d6710ef2e41492b21763b7660073012c0affca72
Former-commit-id: f1c99b51757b3e2976593f9d21767bd352689182
Former-commit-id: 326397a3b6c8cce5867295b58285f8d330ed70fa
Former-commit-id: ff1b3aa64efa3ab7319104f7113e87304c8e1306 [formerly a92d5da50f2faad79ccd5acaf2b803fb56f1b6f4]
Former-commit-id: 2975c03709dd967ae90880d3274af580a96df431