Apply patch to boost
When using:

    scons strict=true ctool=clang

A number of issues appear concerning a known bug in Boost::Bind; applying the patch which should appear in Boost 1.63. Comments on the Boost ticket indicate this may also be an issue for earlier versions of GCC.

For the ticket, see

    https://svn.boost.org/trac/boost/ticket/12397

and the patch is at

    https://github.com/boostorg/bind/commit/3c56630b5400c43d1a4393d685a407e68a69ce9e

NOTE: This does NOT create a dependency upon the patched header. If the header is changed, you must do a full re-build of all sources by hand.

Former-commit-id: 30dc66af4bb87c85676836ec36dff20b662d1981
Former-commit-id: 43634d90c28a9bf25a954cbc6b0af5a83be5add6
Former-commit-id: a14921ea0642efa7384d400a6f9d46c493387fa8
Former-commit-id: 402a3abd83f5ea0fe3127087474d084b91201efa [formerly ddc4b45817fe813d6be6e116c0f486178519ef59]
Former-commit-id: a2548616713717ff46824b673806aa1f43bdbf1e