Micro AIs: correctly deal with hidden and petrified enemies
Previously, the Micro AI behavior was inconsistent at best when it came
to dealing with these units and could even result in AI errors when an
AI unit was ambushed or a petrified unit was in the way of a move. Now,
both types of units are properly "ignored" and the AI moves have been
made robust against unexpected events such as ambushes. Incidentally,
the latter also makes the AI more robust against WML events doing
things the AI cannot know about (such as removing units).


Former-commit-id: 60497c827c1c5cee96eaa75824f3136759c54db0
Former-commit-id: f676a432bee9d167da3c052608a9e38eda8c78d3
Former-commit-id: 22aefdf966ed758700856bdfd4d9667760ead683
Former-commit-id: a6c85b348cf4bc003af20d1798b4ff268dbd2a03 [formerly ea758106b117275f2a11f0a98a8d529b7b0bd470]
Former-commit-id: f9b1199e87d8a89eb6889f720258c2582816e248