Fast Micro AI: simplify code by using new ai.aspect.attacks
Note that this requires that the enemy filter in the utils function now
matches the enemy units to be attacked, while previously it was set to
all enemy units NOT matching the filter. As this is not used anywhere
else, that’s not a problem, it’s just noted here for completeness.


Former-commit-id: 2aa3aa3942e489922c0211f489b3526a8dbd4b2b
Former-commit-id: 55718014a59768ca8e7003baac7c17dfa102037b
Former-commit-id: d431232ecbc2bb3cb23d2a1538bb6667a629b6bd
Former-commit-id: 62d3b88bcc6883266e5d2a0d43c2eb9d2ad41101 [formerly 6dc82cf88d007b42ea902e1740d4a6723361168b]
Former-commit-id: f30cf2d6a4682b352cdc9acf04b5a8b053188475