ai_helper.robust_move_and_attack: return action result to calling function
This function is supposed to be able to deal with all kinds of rare and
weird events without throwing errors, but it should still return a
status table describing what happened. The vast majority of situations
will simply return the move_result tables. For all the strange stuff
that could also happen due to WML events etc., we set up dummy tables
with the same structure and (somewhat) descriptive error messages
instead.


Former-commit-id: a227910e47ddd278a28cbd82373bcd6749316120
Former-commit-id: 8151f2939a720d5c32432cc97505e12c8210741c
Former-commit-id: ffd4adba076f22d3431f9f811aa6cf0cdd348f65
Former-commit-id: 9d2a23d9e4bbbd5eb8c45cb27c2b6b5ce2423b98 [formerly 358251ead34dff385e6b05d90dffa46eaa923a83]
Former-commit-id: fc0cd23be873a1ac806d654f0bf70362f46d5712