MP Create Game: delegate tab content updates to game selection
Previously, on_tab_select was called every time update_details was called (including on game selection), as well as
when a tab was manually selected. I had enabled the updating of tab 3 (game settings) on every call to fix an issue
where game settings were incorrect if that tab had not been selected. However, it was recently discovered that the
same issue affected tab 2 (custom options). Since the only time the contents of the tabs change is when called from
update_details anyway (manually selecting a tab did not), it made sense to move the updating code out of on_tab_select.

This is also a performance optimization, since it prevents tabs' contents from being updated every single time a new
one was selected, when in reality there would be no change unless a game was changed. This will also prevent any future
issues of the same ilk and ensures the state of all widget is always in sync with the settings of the selected game.


Former-commit-id: ad4c65a458ea5c657c14c5a5b4162d01724375de
Former-commit-id: 95970b60767c5e64374c6e2cc2b1347005cc08df
Former-commit-id: 7dd1c7036c20143d3602cb6b2d6172432935c856
Former-commit-id: 55e44efcff89f09dec32541d6d92f446fa917b1b [formerly 6d772b733f2d7d3fa8c6a49209830ebd18180d11]
Former-commit-id: c73625ce0acb977d60294dcd7b6e8a945d63df78