MP Create Game: don't update tab contents when regenerating random maps
This moves the tab update code to on_game_select from update_details, since the latter is called when
regenerating random maps and doing so has no possibility of changing tab contents. The only thing that
can do that is selecting a new game. If this ceases to be true in the future, this can be reverted.

This probably should have been done in 33b653dbdbc5 [formerly c73625ce0acb] [formerly 55e44efcff89 [formerly 6d772b733f2d]] [formerly 7dd1c7036c20] [formerly 95970b60767c] [formerly ad4c65a458ea] - in fact, the commit message makes it sound like
it was, but it was not due to an oversight on my part.


Former-commit-id: 54ac5987b6712913b58b13b149e453312328518a
Former-commit-id: 00729811ccad9e7850eab77e3adf8fc883d27047
Former-commit-id: 90a91ca9e4dea7f8dbee6fc680a3f08f29372cfc
Former-commit-id: 8302a8c9986bd62c020e5dce91b7bbd99dbc5a30 [formerly a20aadf711a9859634b6c1839670391abf7e51aa]
Former-commit-id: fb115e6499ac39960fc093034faa718eb18b2f43