MP Options Helper: skip any game, era, or mod with no [options] subtag
Previously, these would still get included in the final options config and result in a [type] id = tag being saved
in prefs with no [option] tags. When loaded in the constructor, this would pollute the options_data_ map, which would
then in turn pollute the resulting config further, ad infinitum.


Former-commit-id: 6d2b65733096189239c8007104b662dcf25e32f5
Former-commit-id: b9dda046d3dc40aaa0fe7cc426d21bb5fb2344e5
Former-commit-id: b5e34c480a4bf64fca79b0bed13ff7dda5eb30a6
Former-commit-id: b1b0238ca21ed953ae4b54db0d2847128a8ed609 [formerly 7ddfe68c685e477254bb97d814a0b8fb7e2f6421]
Former-commit-id: 4d6625ea3402e270f2d879faeb99c0aaedeaaf13