MP Options Helper: refactor data storage structure
Firstly, this avoids creating duplicate option_source objects in both visible_options and options_data,
and instead makes use of the source id key.

Secondly, it replaces, the sub-map with a config. I don't know if it's any more efficient, but it allows
a simpler transition to option_id = value format after the GUI1 MP screens are dropped.


Former-commit-id: 486e473cc60db973f8583298138e8168861ac441
Former-commit-id: b31702d30f64cda071a4b0c1c872b6e280308be4
Former-commit-id: 00328f622a620063a9ce3a201e46ce1fee3afe19
Former-commit-id: 2e20d62f39b457851bbec5a25dd1f73dee529d1c [formerly db68a9ab84b672b0a42cb01a637fd0923e3dbcec]
Former-commit-id: f85bdd36d13465efffdab042361adcaebf45b836