This removes references to blit_surface and introduces masks for colours
blit_surface is slow compared to the SDL functions and is something we want to remove in order to add hardware rendering.

colours are currently defined by people manually typing out the bit masks, this
lays the groundwork for defining them manually each time. This also starts to acknowledge that alpha is a part of colours.


Former-commit-id: 7b052bbe4e54644d870571d25872fac4aa622aa4
Former-commit-id: 705afc191665fc1bcca3b48dc22ec9763143f6ea
Former-commit-id: 6d11e3bc88c7c76d5ed4378803d59f97436394c2
Former-commit-id: 49ec430302dd5d3825aa61c8d288b9ac12bda415 [formerly 48a14f40ceeaabbc2d0afd1a4c37efe71481f724]
Former-commit-id: 5c7489d23a271695faf45cfc1bbccd98a23692b0