Fix Lua error when WML assigns a real number to an integer attribute
This regressed in the Lua 5.3 upgrade. LuaL_checkinteger() started to give
an error when the variable isn't an integer. In Lua 5.2 it truncated the
variable instead.

Fortunately we have a macro for modifying attributes, and thus I can fix
all the errors at once.

Maybe a version of the macro that gives an error when the attribute isn't
an integer would be useful as well.


Former-commit-id: 2e9b9eb3f3b13bf3a4283cb81e5daddbfd747aee
Former-commit-id: bc8c17ded7419d293b22b8fed92091e83e822ea6
Former-commit-id: 44489a8d4c614fdec44982d776e8a9de478d691a
Former-commit-id: 4a65358340c50d255d6985d86f6cde6fc323f240 [formerly e5c4e39d8ee46d4ebd9c907693e185f76d00ac78]
Former-commit-id: 28d6805cf74575d98440b589b002be8bc98c2839