High-XP Attack CA: do not use AI leader for this type of attack
Otherwise it is too easy to exploit this by baiting the AI leader to
run into a trap, and it is too difficult for what it's worth to make
sure the attack location is safe. If only the side leader is left to do
such an attack, it's more often than not not a good idea to do so
anyway.


Former-commit-id: b070f7fe76df67b3c1ba535c5f864fe9e465b731
Former-commit-id: 76999306bd538dcd349468907ef2d08ead855679
Former-commit-id: 17c97976cf713784b23520ef098f520f88c7534a
Former-commit-id: 3ded57131e5983f878e154fbad14848b1dd5fc6b [formerly f6a0aace4ecb4478f503a256ba0c75f7ea597e2f]
Former-commit-id: bf397aac12f44932d574c51e2c53426dc5f990d0