scons: Exclude Clang from GCC version checks
SCons stupidly treats Clang as if it were GCC (hint: it's not) and that
causes the config check to fail due to the version symbols being
absent/not what we expect. So just ignore the GCC version when Clang is
detected, in the config test program; we can probably expect whoever is
using Clang to be able to deal with any issues that may arise from
attempting to use a Clang version we don't support.

As for why we don't test the Clang version number, according to
<http://clang.llvm.org/docs/LanguageExtensions.html#builtin-macros>,
"marketing version numbers should not be used to check for language
features, as different vendors use different numbering schemes." And
indeed, it is the case with the versions of Clang shipped with XCode
that they do not accurately reflect the upstream version they correspond
to and have weird version numbers such as 5 (as of this writing upstrema
hasn't even released version 4.0 yet). Such is life.


Former-commit-id: 223b8a65ed9351812edeb6a574480472542d4038
Former-commit-id: 89b541f06c05e4c4a4ae9390ec845056aee15ff5
Former-commit-id: 828630a8f250237dacf949afeeef22f21f6a03f4
Former-commit-id: 874f0d6407f24814ab31e2270b1db32ba33553c6 [formerly affb7654bb0fea84861aa77922d937f6ac36d8cc]
Former-commit-id: f931bf4d665e1eff1cdc96c1a3cfb39718e7b0da