MP Create Game: fixed crash when switching to game category with invalid games
This manifested itself when switching to User Maps when the only available map had invalid (campaign-specific)
tiles. Normally, the entry would be removed from the available game types if empty, but the check that does that
only checks for a presence of a game, not validity. This caused a crash in on_filter_change, most likely because
there were no games to apply a filter to.

For now, this fixes the crash. I'll have to add some functionality to prevent starting a game in this or other
situations (such as if the filters filter out all games), since one can still enter a game on that "invalid"
map, or the last selected game if a filter returns 0.

This also means that were one to select a category with a single invalid game, try to start a game, and re-enter
the dialog later, the screen would essentially be blank, as no info would be available to show. I'll have to fix
this post-1.13.6.


Former-commit-id: e49910071f2614b604f41f025224aeec3ca0e3b0
Former-commit-id: c889664c1fd32913215268da1354f802ec29a07a
Former-commit-id: 2d6db94b4305efbfce4a86363bd76ca28fed19cc
Former-commit-id: e431c0b482a16d9e47695d885d7c2873c303aa94 [formerly 332a6f1f1e0e31971872eb0cac9e9504b7ecb474]
Former-commit-id: 1a2f7548f1eb614526e2d0b093fd17ce402c2a72