scons: wesnothd does not need dummy_video.o, breaks wesnoth.org builds
We don't have SDL 2.0.4 on wesnoth.org, so it's not possible to build
dummy_video.o because the config checks fail and SCons doesn't add the
SDL include flags to the environment, and somehow doesn't really warn
that they will be needed for dummy_video.o.

However, since the CMake recipe does not link dummy_video.o into
wesnothd and there doesn't seem to be a reason for wesnothd to require
it (even before we dropped the SDL and SDL_net dependencies from it), we
can probably do away with it.


Former-commit-id: 3cc5279122441ec07f5ad71a7f89f4652d098b98
Former-commit-id: 004e6f2c2570989d5288706981a0960dd92935dc
Former-commit-id: dde292c0eb915259ba8757c708c70b44e21dfa2b
Former-commit-id: 2e50c91581639ea97105aaf47896701b0bcb0c42 [formerly a5f7d1713ebec03b93afa498d5039394cb8f2a40]
Former-commit-id: f119aaac22f5a58bec122b8d2ba627f86bfc5802