Fix compiling with TDM-GCC-64 and -O3
By default, TDM-GCC-64 was using -march=x86-64, which for whatever reason caused issues when compiling with -O3.  This makes scons force it to use -march=pentiumpro, the same as TDM-GCC-32.  This also changes the release build to use -O3 instead of -O2, which matches the release build default used by cmake and code::blocks.

Added -march=pentiumpro on Windows for -O3 build

m5


Former-commit-id: 4c15205f6e0e425f6cf48cece2dca728e20efdee
Former-commit-id: 0da924d677694734889e59501ff3d1d25141c113
Former-commit-id: bf928e532b0be24ba72c9b1c5f2cc07cfb6de12b
Former-commit-id: 95ea882aac465d4490e3a23da5b35bc8dd1471e6 [formerly 7d322c1c74d606c0d993b2f666a705a8df1314e0]
Former-commit-id: cfddc2d26651287046ccd82084dd83ecaccf0923