Game Config: fixed red_to_green and blue_to_white returning incorrect values
The issue was that load_config is (for some reason) called twice, and the respective vectors of colors
were simply added on to each time, resulting in incorrect color values when queried.

I figured it would be slightly more efficient to simply assign the vector a new value since the size remains the same.
I'm not sure if a simple vec.clear() call would be more or as efficient, though, not that it really matters at this size.


Former-commit-id: 50d0219016ea16e187c08074c91855d79bb2cee6
Former-commit-id: 630173b2cb44141ed4f901fdac394a5d69dedc41
Former-commit-id: f6cb7dbdd708c7ec8047e6fbba77ae72e69e745b
Former-commit-id: 552bcc238383c5c5bf2e787ea9ad39588b1473bb [formerly b71f8a1410e0702e9d77235d4183249ee2d538d4]
Former-commit-id: 7d49057bfef7f7e7c1117e9d5f98dbcc60c27df2