deprecate needs_select=yes
The behaviour of needs_select=yes event is rather strange and hard to understand:
When a menu items with needs_select=yes it clicked, it first fires the previous select event again but this time in a synced context, so that the select event was fires 2 times total on the playing client and one time at the other clients. It also the breaks the specification in the wiki that select events are not fired in synced contexts.

Also it might happen that the unit stands no longer at the position from where it was selected because it has moved causing unexpected behaviour in this case.

The new way to let other clients know the currents selected unit is the [sync_variable] tag which was added in 1.13.

Former-commit-id: 9cc9cfd1d56c0b69e1c782c8fcea2661e6802180
Former-commit-id: d2d9ea7020d8620194f73d2e832eb226759ea34e
Former-commit-id: e4a750064c60dbbd9b7eab3a440c751e4707f6e0
Former-commit-id: c478d607250452dc48d0f6ad11cdad08af941f94 [formerly c66c2c5d0299a3ce6ea12a5c8850534425d4e8fb]
Former-commit-id: b2e95160fd073b02a3199a0c91d7aa9ec44caf70