Stop replacing listboxes with panes with --new-widgets
I don't see the point of such a change. More important, it breaks all
dialogs which search a listbox with find_widget<listbox>() and don't have
a separate code path for --new-widgets, such as the new add-on manager.
In particular, testing the new add-on manager without code changes was
impossible because it was disabled without --new-widgets and failed to
find the add-on list with it.


Former-commit-id: 6710cdb9daeee6ce033706c4d391020ab57de86c
Former-commit-id: 7094b01b26dd9ed3ad5de0d6a6ff2152efc95453
Former-commit-id: 64599bd5656c6e11117f4502ea1cf04f1b29c310
Former-commit-id: 1a073b7664910a0e36e9345320d53236b91b2832 [formerly 761c26591e3726d34079cf6b9af11034c2a5c8fc]
Former-commit-id: f52ad28da51cee3841c6505174c1f752e358d62e