WIP: make the add-on list a widget
The purpose of this change is ability to create more dialogs which present
a list of add-ons, such as a GUI2 version of the "Install Dependencies"
dialog.

This change is still highly incomplete. First of all, I disabled inline
install buttons and ability to sort the add-on list. In addition, a lot of
functionality that belongs in the add-on list widget is still in the GUI2
add-on manager instead. It is also notable that the widget relies on
several linked groups being defined in the window.

I plan to finish the transition to the add-on list widget, and address the
problems in the previous paragraph, in the future. The linked group issue
is going to require a lot of work in particular: I plan to either make
linked groups grid-specific instead of window-specific (like they are now)
or allow widgets to inject linked groups to the containing window. Either
way, I'm going to change the operation of GUI2 fairly deeply.


Former-commit-id: 347d177f805be6d7a291ad79633784acec2e8fcd
Former-commit-id: d6c8ac0915af7a7e3f4d8f9da76ff4fe8676db1d
Former-commit-id: 44049b3a1bffbe081d8a076c186988bf21bd67db
Former-commit-id: 761e3f9e98167f3b386c54ad28eed3814a946f4a [formerly 971e7efa85532deffc7efc5eb31c5f77b49ca250]
Former-commit-id: c25d810fe45c295bcfb82b6e814979f451ceb9cc