Drop Down Menu: several improvements to embedded toggle button handling
Any click on a row with a toggle button will now toggle that row's button prior to closing the dialog, unless the
keep_open flag was set, in which case the click will toggle the button but the dialog will not close.

Clicks directly on the toggle buttons themselves are unchanged.


Former-commit-id: 0891a1b3b439cfe31b6aff4c8c9d24c8af807e3e
Former-commit-id: 2b7c50444e3022cbf6fbfe0faf93d1e60ea7dafa
Former-commit-id: 539f12599dc9f895df4149398cf52aa5ee1684ff
Former-commit-id: d96bc668608fc2cfb4f73fc9ddf6485acf2098a0 [formerly 2dd8ff41ddcaf643a3a1745fd7cbd698838648d8]
Former-commit-id: 4ac918085741906b1c089cfd7fab54eb21e24973