Attempt to fix bugs #25524 and #25525
In order to fix use_map_settings not being set, we create a configure_engine object in the sp game setup
and pass it to the sp options dialog.

I also added checks for the presence of [options] tags in [campaign], as well as check to ensure empty tags
don't cause the dialog to show.

@gfgtdf please confirm fix


Former-commit-id: 2985101148f30f843e9338328b0b5bcd197b1df7
Former-commit-id: 8e0b228aca3809f1aa16fb58b0013e8d3f5d3242
Former-commit-id: 8d124bcde726dc8e6433d180365313a04ad2c8d6
Former-commit-id: 8d4cddd05e9b2a1ee858dd23a195224dc7ee064b [formerly 13624f4819e31411c40e971f33d3f5ff6f739941]
Former-commit-id: 8325e3a7a31cc8f7c8070b24bff408c4817c5576