Campaign Dialog: fix mod toggling (fixup 66b590b7)
Turns out I made a mistake in that commit. Since the create engine interface relies on toggling a mod, not actually
setting its state directly, my method of toggling a mod if it was selected meant a mod could never been de-selected.

This reverts mod toggling to a callback-based system, using the method implemented in 693e4b4f.


Former-commit-id: aca354e609862368fb5379331831c28e644d9a53
Former-commit-id: cd59a026a986ffc7ddc4e2d2c1bb13861db7b6f1
Former-commit-id: e74d5531170a95020cfda29f4726d13164a8fb8f
Former-commit-id: 227d2d092d8bd82eaf0270bbeeb74814bc4d9442 [formerly 4407e875a8b0659a16faa43bc8762cbd09b22294]
Former-commit-id: 49457deb12e8ad7e7306f8037acffd2c705d787b