SotA S2: Fix exploit allowing WCs to be created on invalid hexes
The leader should be able to raise corpses only on adjacent hexes.
However, it was possible for the player to right-click on an adjacent
hex then quickly right-click on any other hex, and the corpse would be
raised there instead.

Allowing user interaction to affect a command while it is being
executed may be unintended behavior, and if the engine is changed to
prevent it, this commit can be reverted.


Former-commit-id: 2dec12deb43c02f655c06101454b2fc020023c31
Former-commit-id: ea54a259f9b9ca9affa7c811c7294589a1d959a4
Former-commit-id: 3a8989544f438ca3a6b0d1f2bbc64120a6a33810
Former-commit-id: 58ae4032d3346722ec6aa8c0065fe80ae7868b41 [formerly 5c621312f70834a38b62757e1196923dd1763755]
Former-commit-id: f2d7b0bb29d5fa704573db9d6571bbde27a50b0c