Fix bug #25538: Monte Carlo damage prediction relies on UB
IEEE 754 specifies that dividing a positive number with zero yields
positive infinity, but the C++ language doesn't guarantee IEEE 754
semantics. As a result, UBSan complains about floating-point division by
zero. Wesnoth requires IEEE 754 anyway, but it's still good to get rid of
the UBSan warning.

The code now avoids the division if the divisor is zero.


Former-commit-id: d62ae964f769e7d0d77bfdd7558849f1c4929385
Former-commit-id: 7d9c24afb6a2d04151d95bd62ac4bf8468b64b0b
Former-commit-id: 99440af41a2af74c748eeb28fe33c959faad20e3
Former-commit-id: fd605640b035d8fb330bfe52faa68916dccee717 [formerly 02819a790550c2bb4e592f704db5b89e72854505]
Former-commit-id: e0a9a614f991c15c3fed5e5bd1425b7752d23d5e