Small fixup to 42f087abb
The old `tristate_to_bool` function returned the second argument if the value was `t_unset`. boost::tribool can be converted to bool,
but the intermediate state returns false. In this case, the only usecase of `tristate_to_bool` had true as a second argument, and I had
forgotten to explicitly check for the intermediate state when I switched to boost::tribool.

It might be useful to re-add the `tristate_to_bool` function again if conversions to bool become more common, but for now this simple
check works.


Former-commit-id: fdeb90ba3b57aaf5144963c0a85636f50fa09d98
Former-commit-id: 7573f7d5568a58594182f71d5cc54bdd6595a330
Former-commit-id: 10678547f23aeb87e54f2f170674c6585d8b69c6
Former-commit-id: ce0d1f12ff56d594e1fb50811b79a2c036f7c9d9 [formerly 63f164654eafda71ee6abf73119ff9db20763d39]
Former-commit-id: a685f2cf49511fa0657ef5ff187d4dc39dfcf40e