Fix some [modify_ai] isses (addresses GNA25558)
As far as I can tell, the most likely cause of the bug was that several
cases of [modify_ai]action=add were called with a path that lacked the final
empty braces at the end of the path.

This commit makes both the tag and the underlying Lua API calls explicitly
handle this case as if the final empty braces were, in fact, present.


Former-commit-id: 119545fce8118a9603e26944479258c3c70d765e
Former-commit-id: 4b26dda81b86371969ffce9d6f56f285562a5b35
Former-commit-id: e5c50bafe9ace0c30e00d83737926c476dff046e
Former-commit-id: c2ab4947437675778a979b16de8ef892e77cf56d [formerly 13f46a325ab80036608414795d1c7e06cb1ab8c8]
Former-commit-id: 5946a3a5be666a47c5472621de4dd7983e7cbdcf