CMake: demote failure to compile IEEE 754 test from error to warning
Apparently due to a CMake bug, some (but not all) CMake versions attempt to
compile the test with default flags, which means C++98 mode with GCC < 6.
As a result, compiling the test fails, and the whole game fails to build.

@gfgtdf pointed out to me that CMake reports the compile and run status
separately, and therefore it's possible to detect the situation where the
test fails to compile. Thus, I decided to allow building in that situation.


Former-commit-id: 6d632836f21903e949e4002250469e4827bfc4c0
Former-commit-id: dd07cfc6c0b66139566e44c81105f3dbc43026e2
Former-commit-id: ca15525e7323da01a70478170dff66acc0999aa2
Former-commit-id: dec070dd4afef5a61f7183aa4d359ae844a32771 [formerly f186fd2edc8a9583985efa531ebe86d23e81d09b]
Former-commit-id: d4e3841b356df277054d571596ab51136ae660f6