MP Lobby: use exit hook for quit confirmation
By design, this extends the confirmation popup to exiting the lobby with the Log Out button too (previously, only
the Esc key triggered it. However, given the proximity of the button to other control buttons, it's not inconceivable
that someone might click it by accident, so the confirmation is useful.


Former-commit-id: ef74b4bd5f40f7301af6ebc1bd3a202feaf40d0f
Former-commit-id: dc56b317b04f4a8d11942c369e681ff5cfe03e76
Former-commit-id: a2ac4674e105b6fef52abdb64d129e7db6a94cb6
Former-commit-id: 7273b434fe6ebff9849c51edd4b4b25ce1a3e1b9 [formerly 75452b21e4ae9c6d05c1be7e89e954804c0dc970]
Former-commit-id: cc7320236c2a8ef87e53eea2147304f77e8c9118