SotA 16: Fix Archmage having move at start of turn 5
Side X Turn Y occurs before the side's refresh, but the mechanics of refreshing don't occur on turn 1.
This was why the first Archmage had no move but the second one did.

Also adjust positioning so the Archmage is more likely to go to attack the player instead of going for villages or acting confused.


Former-commit-id: 11bf1cb2dd06bd59eddaa410fa8d3b20750e2fc9
Former-commit-id: 9c7aebe3555933789d928cae0f4336b9c5e19efe
Former-commit-id: 2b2d230e3403f47157ff9aae1de20dfc64e4e986
Former-commit-id: 968692c8751867023e0d601b1815a3434e082e20 [formerly c969b227d56c2b66b39fcc424696752ca2aa8ec8]
Former-commit-id: fab9824b2b57c538d52aa1f198ac3d2b409499c6