Fix "failed to place a grid" messages when partial relayout is used
The problem was that some widgets which wrap a grid didn't have their
cached sizes thrown away when the grid itself did. As a result, when the
parent grid of such a widget tried partial relayout, it discovered that the
widget fits just fine (according to the cached size) and attempted to lay
it out without shrinking it first. Then the grid found that it was given
much less space than it would prefer, and printed the error message.

Fixed by discarding cached sizes of such widgets.


Former-commit-id: 5e65fa1d59678cd958a0bd46b652fa4d6f57342d
Former-commit-id: 25074bf7335fafb46c3c00dcdea7042021365554
Former-commit-id: 882206e759338e509ca52de7242dd39ff7067ce0
Former-commit-id: f8c4f1de3927cf0404bf5f11448d579bfa4af2a0 [formerly 48f1076851116172e9f23e2fa3675f8b1a6a3a03]
Former-commit-id: f3585c9df8693ac00a75cac1141c09751e2cb7de