Refactored handling of the callable variant debug stack
The debug stack is only relevant to callable variants. So instead of continuously passing around a stack reference
to all value classes just so it can be passed back to variant::to_debug_string in the case of recursive calls,
we now keep a static stack member in variant_callable. A second bool flag (clear_stack) has been added to
variant::to_debug_string, to which all recursive value callers should pass false in order to preserve the stack.

Former-commit-id: 108e4322d51603a3e45c19bc367800945399b3d5
Former-commit-id: a6a8c70e5a0e3b2ada4cdd3427999f3fc4195ba9
Former-commit-id: c72f0b07cb108517453b72f2e72fde9f42aefe84
Former-commit-id: a3292fc0d7bc21684e188ad35d3162177845c5b0 [formerly f0933b3e5e49e95c2183c7865ccebbd2abcb2460]
Former-commit-id: 4fad0673ed843ad7f1773d6b6b0879b74bb304ca