Swap the arguments to variant::to_debug_string
The seen stack is an implementation detail that should generally only be
non-null when called recursively from the implementation, so it makes
more sense for it to be the second argument rather than the first.

I did not swap the arguments for variant_value_base::get_debug_string,
because that's entirely internal anyway so it doesn't really matter.


Former-commit-id: ceb2cd5c8d9cb981c499ea3be57522e3a534955f
Former-commit-id: 516ac03d17cdbd1951f088a788c3300c0d2f44ca
Former-commit-id: a079900574d92eaa99af34fbf42eee46de4772ac
Former-commit-id: 40cc463301bb7dcd9df67aab36ce49d55aa1a9ea [formerly 671fac5ce450a6fb84ee5d75c40da62169fcc551]
Former-commit-id: b6ecf24e0bb505fc35b622ce0709d23f460ed56e