Story Viewer: deployed new alignment set API
The set_origin method worked, but not in every circumstance. For example, if the story box were not aligned to the bottom of the
screen in the initial part, it would not be correctly place. If you resized the screen while the box was not at the bottom, it
would return there. This was due to set_origin being purely temporary - ie, the layout engine did not record the new position,
and when invalidate_layout was called, the default (per the dialog WML) position of 'bottom' was restored.

This new method fixes the above issues. It also fixes an issue the old GUI2 screen had where a top-aligned box would also obscure
any available title.


Former-commit-id: 8260f4050dfc29a6fb3aeef5bb0228acb4d789a4
Former-commit-id: e894f74678b4cbd9eb4142e85b797b0b9c267c10
Former-commit-id: b2bfb9eab2cc2b86175b3ad40634066c99c32425
Former-commit-id: 41f231ec44953e2857b7e2b06ece2664bb5d4bbe [formerly b87fbbb10b8cebb1d4cd25fa7335b4d76e90f882]
Former-commit-id: 061039876cbc3fcfb364c48ee3c0da51607f88ee