Story Viewer: fixed all floating images drawing in the same spot
Since all the floating images' data was set with a formula that relied on a canvas-wide variable whose value
was constantly changing and no serialization was done, the formulas all evaluated to the same location.

To fix this, we do away with the use of the variables and insert the necessary values directly into the formula
string.


Former-commit-id: 9a27e3b2f53aae6b25523958f378b9ad88028d4b
Former-commit-id: 3598b6297800ce05f8f0835feb2fd8520fd70c3f
Former-commit-id: 575d5afb947f4989ed1d91fe26cc581ca72373d3
Former-commit-id: 622ac4ab9591e5eba45095fbcb8f9f5de753ae22 [formerly 651a0be7692a7168767c6ba009a62985969f2e5b]
Former-commit-id: 83b429002f0256005731f7c17d2996b26a23252c