Refactor handling of the ToD manager
Instead of the individual classes such as game_board or map_context owning a manager, we instead use a singleton-like
design where the manager is reset or cleared by individual classes, such as the two above, during their lifetime.

This also removes resources::tod_manager in favor of the singleton.


Former-commit-id: 2dde497760c6eec9c30574c3e75b575870ea90bf
Former-commit-id: e2941737ee5f82d9705f55ff6ded4e91cf74cd1f
Former-commit-id: 74401b52da14463fa2ea71a6a94adc0e90b61530
Former-commit-id: e5f433e7515a05652386b3607a91a5cd4618c542 [formerly b671d8e800dc3bf5186711bce08333c180db1311]
Former-commit-id: b12b739d9fce27a94186507ddaba8f70ed6948c8