Display: removed dummy drawing layer
Not exactly sure what it meant by "sizing the vector" - the enum isn't a vector, or used in a vector. It doesn't
seem to have been relevant to max_layer_group either, since that was calculated with a -2 offset to ignore it.

Likely it had to do with the loop in drawing_buffer_key ctor, which was designed to find the group (which is
drawing_layer an enum member) "under" (ie, that contained) the specified layer. Having a dummy buffer meant the
precondition of layer < max layer (ie, the buffer) layer was always true, meaning the index was always decremented
at least once.


Former-commit-id: 03981acb89a40559eedac14b6a8b90a0f9475f09
Former-commit-id: 07478c41c02e5ece2fa0e73a70aa25a08094e291
Former-commit-id: 8d8ba73230bf8e0912e4c6f3eac2319a7276141d
Former-commit-id: b2cc0d5d25b0e03a218d323d3c45a7ed4b0ea9bb [formerly 542750b4d2624ae3ad4912530b5529570f5bdef8]
Former-commit-id: 290efc4d657823895ef5e91baa7cca2027acc223