Revert an unnecessary part of 7d2d100
I realized the the theme menu class doesn't contain the *parsed* data - which should indeed be in a config -
it only contains the *ids* of the menu's contents. Therefor, there's no need to store the list of ids in a config,
since it would only have one entry. A string is therefor sufficient.

To enable the input and output vectors being different types, I cleaned up the output initialization loop in
controller_base::show_menu


Former-commit-id: ae0a60a2096472001bb642c46d759ef06695eb0a
Former-commit-id: 21e21f6e0917e1c996740b1d47cf971fa41c2d3d
Former-commit-id: 45283ae3fbcd82efe9c9f67ff03f23bfb79dfdf3
Former-commit-id: 76d42d8966caed6d5af97b3ede668753fc40dbc2 [formerly e27f897956e0a5f8c78708302ae2ae027421830e]
Former-commit-id: 372f94949d723f17f91a6f274b2dd279a26fa62c