Cleaned up menu-related code in play_controller::hotkey_handler
For show_menu, there's no need for an explicit check to "save" the autosaves entry anymore, since we add
entries to the item list instead of removing them.

For expand_autosaves, I shouldn't have made the saves list a config just to transform it back to a string.
I only did that since trim_items took a vector of configs. I've now made it templated.


Former-commit-id: d8d7e82c740d554390728a1c1f3aa099aa771b19
Former-commit-id: 65d88956dcf1456ceac9b2bb4ef5223077a0c0b7
Former-commit-id: 81f79ce7898b09e30f8e033212516db4d1b4450d
Former-commit-id: 832ce6a2efead87e68cdfa848dcd7c93b829e6bd [formerly 0f79c835b5fa4995dd80291b5b00cd31c45ca3f1]
Former-commit-id: 6d613bfa892275f384cbad924ed8246c04d94276