Editor: fix crash when closing map (fixup bf79dbd8)
I had moved the creation of the map_context_refresher prior to deleting a context since I thought it was
necessary to not access invalid memory. Turns out doing that was wrong; doing so caused the crash.

However, this results in a slight change in the code's result: Before, it was taking a ptr to the
current context, modifying the container, and then creating the refresher. This allowed the size_changed_
variable to be correctly initialized. That variable is unused, though, so it makes no difference.


Former-commit-id: 11e53c88992e41f756b45568d4063d5dfe7d916e
Former-commit-id: 92e40fc1763b6d4a61b2d8d765452d6ffe8ee1d1
Former-commit-id: a546d1adfede56363fc7233f495071b1e8d53755
Former-commit-id: 7c55eb8ff71061231cbdc1dfc9427465f3c69e59 [formerly d16a1da7f748998f95d9065ca71e94f04cf8d5d3]
Former-commit-id: 0362fe8492164ac35879c8f7d8dc2aa917556963