Pipe raw SDL events throught the GUI2 event system for hotkeys
This adds support for passing raw SDL events throught the GUI2
event system in order to allow for hotkey creation in the GUI2 hotkey
dialog. The dialog in turn uses the raw SDL event to query the hotkey
subsystem whether it's a suitable event to create a hotkey from or not.

The first suitable event is used to create a hotkey.


Former-commit-id: a2a0bf9119eba87b2abe148454333eaf2f20f2d6
Former-commit-id: cb438742e2b2d9da98ca896f6f9e419520cd50eb
Former-commit-id: 85721afaba89ce86a178d0ed9d0dd9a303a40773
Former-commit-id: 86b30041c96d455c4315dcda4c04c7fe8aa7b545 [formerly 8814e6dcf59378695253ce494b213dcd41196b1a]
Former-commit-id: ce82433b1f8eaa17aea1aec7830dfd39b7a34874