Story Viewer: delegate storyscreen controller creation to dialog
I left the controller passed in an argument for compatibility with the GUI1 dialog, but since that's removed
it makes sense that the dialog handle the creation of the controller, especially since due to an oversight
it made a copy instead of holding a reference.

This also allows storyscreen/interface.*pp to be removed. The START_POSITION was part of the GUI1 implementation
and isn't needed anymore.


Former-commit-id: 010fc0192ce5a544eec4a82b33968be5d6054b5f
Former-commit-id: 3919c7cba05dd96e3dfe43c271629fe4fe0a91c8
Former-commit-id: 897075f80bf5299a412666a7aa624d9c4778ed3e
Former-commit-id: 9989fc5fc784d22ea309a0cf03ee2ecafa08893e [formerly c49bb364d4f818a2e1dded88cd5a9d8ea6368695]
Former-commit-id: fdc8ed641cc1387caca4ccffc95bbfb98ba0b34c