Editor: refactored global labels handling to avoid static pointer to context internals
This should *finally* fix the crash when closing a map (bug #25680). I forgot that the pointer
wouldn't automatically be reset to nullptr when the context to whose labels the pointer was to
was destroyed. Would have been simplest to do that manually, but this makes the code a lot less
confusing and avoid future scenarios where that pointer might be pointing to freed memory.


Former-commit-id: 766d5a71cb3aea123c47e1e6e087892d423f78c5
Former-commit-id: 0f4444bdcc1745affe1ff9fcdbf39b59d32001d1
Former-commit-id: 53d62a073ccd62f2a298c13e217e30f38901e304
Former-commit-id: aeaf832f94633257002624eab53a0e408695c6d5 [formerly 841682987d413813f423c83a3803e89cec1cd92b]
Former-commit-id: b37d96aa2bdee5debb705d8aaac1933f48dea280