GUI2: removed unused Campaign Settings dialog
This was a dialog used early in 1.13 during the SP/MP codepath unification to configure certain options.
The modification selection is now built into the Selection dialog, and we decided not to enable other options
like Era selection in SP. The only option we might look to restoring in the future is the "show_connect"
option, but that requires some additional code elsewhere and probably won't be used anyway.


Former-commit-id: 9ad7c188ec76f9c7946f4ca927ce1abd74cffd23
Former-commit-id: 0b68e7dd3e78a9149d7d7f361d96aea28331bea1
Former-commit-id: 7f96f2424337d3faabc44e0d8b71305efe9766fa
Former-commit-id: 695c9e892aaa5674e093503d54152b7950bfd87c [formerly 6eb6f898a8e3c91af0ee6358c317a90af0caa5f7]
Former-commit-id: 2e854abafb549094e11beccd8135aa5945f4bff3