Editor: fix button overlay disappearing when a GUI2 dialog was open
The problem was that the redraw observer, which handled setting the button states - including the overlays, which
aren't specified in the ThemeWML - was called *after* a draw happened, meaning the buttons were drawn without
overlays properly registered.

This moves the firing of redraw observes before the draw call. The only code that used the redraw ovserver functionality
was the editor, so this shouldn't break anything.


Former-commit-id: dfccede885eaf6554b8f3476cd071ae23b3aed45
Former-commit-id: f1f6c9c8c0e73fce67779d4111eedf4126e0453e
Former-commit-id: 49e0c90bde44c9bbfcc04851bf82f2ba7d6ed561
Former-commit-id: 216a224efddec47ed4fe729886999aef7f339b6f [formerly 24cd1f8085777c83a907283cf1f24dcd4901ed38]
Former-commit-id: 5b7d27fc5df5546bcfdb2a0a0138246cb8ff70a6