Editor: 'unhide' palette when drawing contents
I'm not sure why, but this allows the enabled state of the to be set properly when switching to a tool
with an empty palette. It seems this function is called after editor_toolkit::hotkey_set_mouse_action
is fired, which already seems to do that, so I'm not sure why it doesn't take effect there, but meh.

Also includes a fixup for when palette_menu_button is fetched. It's a menu button not an action button,
but it probably not havig worked before seems not to have affected anything.


Former-commit-id: 958477857dea973ca2672c7c2603f9a822eba718
Former-commit-id: 7040de5c749e698b783afc974eec06a1bc3f1755
Former-commit-id: 251a8ea7d2d3e82dc67dc24f25be141bbf6137b0
Former-commit-id: bae673e6cbdb0ce075c028bacbb06f641772caf0 [formerly f26217caeeff8131ae1417084d7507ade042949c]
Former-commit-id: 1c0cb5c50d495e17af34f4517c97d907c8be5b47