Reduce invisible units' transparency (#987)
Units are rendered at 0.5 alpha when they are invisible (for example, Elvish Avengers' Ambush ability). An issue arises with the Shadow (and Nightgaunt) ghost unit which has the Nightstalker ability allowing it to be invisible at night time - in combination with its fluctuating alpha animation, it can be quite hard for the player to see at night time (https://gna.org/bugs/?14503). So this change slightly increases the alpha level of invisible units from 0.5 to 0.6.

Former-commit-id: 3840af67cbc4f3dc3d95741777ee87dc34d922a7
Former-commit-id: bb1d9dc16d2175674a37a6ce5f57c961d5470431
Former-commit-id: db6fae4824321a58e89f24b3df09296ea5bcd3d6
Former-commit-id: 92a042310fe548c31966a904a629294af573c050 [formerly 31b858d8bff2a48674858e4bcc130a94cca9c941]
Former-commit-id: 941539f92571455dbbcd99c13850515e11da1f3a