Fix window stack for modeless dialogs
They were not being added to the stack at all, because unlike with
modal dialogs, the window is still visible by the time the show()
function returns. They should have been removed not before reurning
from show() but instead in the hide() function.

This commit only does this for non-tooltip modeless dialogs - dialogs
that allow interaction. The reason for this is that such dialogs could
potentially contain a textbox, which uses the window stack to determine
whether to show a blinking cursor.


Former-commit-id: e73ed5703f416bd6261af71e88586106c6c896b8
Former-commit-id: b187b0b0ac281ff2b2f734fff8006a72ea1ec99d
Former-commit-id: 8b175f5306021f4cc6284e8439c1384046e7cdaf
Former-commit-id: 38ab47c0c6336c339a1416f9e0df841ac694a961 [formerly 3c19e4ae52de64e0ed76498bfde9e810bb523339]
Former-commit-id: 6a73f78c782ee8e4b77d12c015751d31e3794c5b