WMI Container: removed iterator and container interface
This probably existed to allow simple range-for code in pre-C++11 days when we were using the
BOOST_FOREACH macro. IIRC that required an iterator. However, with C++11 and range-for, it's
much simple to simply specify the `.second` when necessary than having a whole convoluted interface
and iterator extension just to avoid having to type that.

This also changes the return type of 'erase' to bool. It also leaves the empty() and size() functions,
since those might be useful.


Former-commit-id: 8fdf6315666e11e4bce3bd4cd5370021326ee509
Former-commit-id: 933203d3d98b1092d611e50e1b83d2131fee8f47
Former-commit-id: 2de2423607049fb34fe690fe8ba1f7cce3191d72
Former-commit-id: f09d567d6ea813f9ec986cc9f053d558e1072918 [formerly 683367de1b8e53e8eba54cc64e32e38c68853597]
Former-commit-id: 6a7564b538fe8d928f72ceaaf1ddbcef21ab1c54