Game Events: throw out custom handler_list implementation
This replaces it with the standard std::list interface. I've attempted to make this change as least invasive
as possible, so I have not touched the manager::iteration class. It can probably be refactored further later,
but the crux of this change is simply to allow the removal of the custom smart_list container while keeping
the chance of breakage to a minimum.

The new implementation cannot be an std::forward_list since that only has front-insertion capabilities
while back-insertion is needed.


Former-commit-id: 6a2c5e3ca2ffdcb6aa14b017af9f866698cae108
Former-commit-id: ca383049d4f8fccde3ff1e3f30a9d995a93332ba
Former-commit-id: f1e714a368485493c5cef63e502bcf1264b44357
Former-commit-id: 2b7c6a84fc9df3edfed040568ffb5b584a9a30f8 [formerly 4ff74224cb4cc9a3d90fdf980a17d7e9aa58b49b]
Former-commit-id: c09a6f64cbae0400a55b98c04c4b4fa581e22147