Add new --report/-R switch to generate an info dump for bug reports
This is intended to be used instead of the Game Version dialog's
clipboard text report in the event that Wesnoth cannot reach the
titlescreen or the dialog for some reason.

Unlike the --version info dump, it performs various calls to the
filesystem API which have potential side-effects on the environment
(namely, creating a user data directory structure if none existed before
running the command). It seems highly improper for a program's --version
switch to perform any actions other than printing text to its parent
terminal, hence I thought it would be better to add a new switch with a
more active phrasing, also reflected in its accompanying documentation.

Ultimately, this and its UI counterpart should be used by 1.14 players
as part of the bug reporting process, preferably as a mandatory step
before submitting new reports on GH or the forums.


Former-commit-id: e5b19c42a384f37d78abadf80bed95f073514c00
Former-commit-id: 98fc7ee0cf971981696ce43fced563436b1aa1f7
Former-commit-id: 66f7fa990ab89aaafbe0cd47417edb58df7e1dd7
Former-commit-id: 1a872b2e0b7a19795dcbc39fa7679eae22c05d5f [formerly 92c60e34a521364cd2816d284c1728e0bd57019f]
Former-commit-id: c7e476d80acf35a4e8da0d3c636822218815b87c