bi: Add OpenSSL/libcrypto version info
Best 90 minutes I've ever spent in my life. I could've done string
comparisons, but somehow I suspect their format is more likely to change
in the future than the numeric versions.

Also, beware that the upstream documentation (both on their wiki and
their source) does not accurately reflect history -- the version numbers
for releases between 0.9.3 and 0.9.5 were vastly different, and this
version of the algorithm I wrote seems to be the best fit for all of
them.)


Former-commit-id: ddb1c5706f6a0b29f0da0db52955179750ff1b45
Former-commit-id: ee95be5861c3613c2c6a72f2a6b99d5fd52250eb
Former-commit-id: 275c067233667ea51b906df9c6b5d5f3a367714f
Former-commit-id: bf32f19e04c144fe94379703f627e302b3d3049d [formerly 001535ca5e5dfffd9a9f99e3fd92b90bcc1cf727]
Former-commit-id: e5ca56387d093a29e9ffb1b5aaa5f3c97035713d