gui2/game_version: Try to not overflow tab selector space
The previous commit adding a new library to the Library tab causes a
vertical scrollbar to appear there and somehow make the tab selector
overflow and get a vertical scrollbar. Vultraz suggested this fix and
seems to be aware that it doesn't solve preexisting issues with overlong
tab labels/translations.


Former-commit-id: 044d30e8ded28b03179bfb9c39af505a6e06f8d9
Former-commit-id: a729b69fe7e1689617d8f376ca8542ad8739287c
Former-commit-id: 8de5e4d94c3c00cc22340610fab944ee47e9be85
Former-commit-id: 4f0ee5798afa6e32c8cff5eda940f3b6cdf65204 [formerly 37cbbdbe5ddef883a96826d58a00a941deb4179e]
Former-commit-id: ae01ef44a8fe529adfcf9157fb7ce98896b71c21