GUI2/Styled Widget: use move schematics to greatly reduce the amount of canvas copying
I added some debugging aids, and found that prior to this change, a total of *807* canvas objects
were created by the time the titlescreen was reached, and a new one almost (I think) every time a
tooltip was created.

With this change, the number of new canvas objects created as fallen to *476* by the same point
(a *331* decrease), and no new canvases are created when triggering a tooltip.

I'm not exactly sure how expensive creating a canvas is, though again, it's not a trivial object.
Avoiding hundreds of extra ones should result in some performance boost, at least.


Former-commit-id: 6981ea43cfbd83eb3e818c9f3e251be5b4f50586
Former-commit-id: b98e45e64b3741f673643938ed9522ae176febb5
Former-commit-id: bcbfdf5a8008170cbe861ec015f4162690a257c3
Former-commit-id: 31a3de1f8cdf2e95c93f226c3e045d0e68e67ce1 [formerly 98c92eb12a64eb24e95781794fb31fddd8676349]
Former-commit-id: fbf1e581cefeaf43629f730d53aa90a3eb9c9ce5