Fixup 9faf17747
It seems while my math in the original commit was correct (both usecases - overflow and amount-to-advance - would
yield in the correct positive amount) I failed to consider the fact that some of the code relied on non-positive
numbers for greater-than comparisons.

To compensate, I've split the differential function in two and made each return 0 instead of negative amounts.
This makes the code clearer while not actually breaking anything.


Former-commit-id: a0e550a61900b53101802f697deb7000ce9366e3
Former-commit-id: a65d77317ce01afa076ee697641c3faa8bbb7774
Former-commit-id: 4b1c485448eced8d048d6b68453725e7b2cc6a2e
Former-commit-id: f772730ad2a914167c9b000c86bc386690fdcaba [formerly 54303ffa4ab46912f4ff9fc78c7b009b69303083]
Former-commit-id: 7256562ba2db8ac2756da0db72888aaa18ac5d33