GUI2: refactored viewport setting and re-added clip rect setting
It turns out I had removed the clip rect setting in error; it prevented items in, say, a listbox from drawing
outside the widget's boundaries. I've re-added that now.

I've also moved viewport setting to the same place. It turns out it didn't need to be set every canvas draw.
The only reason the old code was passing the blitting rect to the canvas was it was needed for the sdl_blit call.
Since I can now set the viewport independent of a canvas state, there's no longer a need to set the viewport in
the canvas.


Former-commit-id: efb7ef91e3edd2ccdd28e11cb51e3dea75d976c0
Former-commit-id: cb987a22950f3fd6691d9502ab0d2933c4ff9fcd
Former-commit-id: d07aca57f5a7178ee1735ac6b31dc11cb420a5b6
Former-commit-id: ad6eb1d1ff8c16f1ca5f4f108c09aab798cc464f [formerly 31e7e25a2c465a0a36004e54056ae13e8dae69e5]
Former-commit-id: 5ddb57559f2b0e26ba9112edc6b347bcfd53179e