Fixup 7fab421f6825 [formerly 9d8a006471f1] [formerly 0fc7df4d5e39 [formerly 4fe4d42e1a62]] [formerly a589ab259052] [formerly 4daff3406b5c] [formerly 97b4536fbce0]
The current resolution values are the effective screen resolution, not the physical ones. So, we want
to exclude resolutions greater than the effective res * the DPI scale factor (ie, the physical resolution),
not the res option * the DPI scale factor. That appears to be because the list of res options is of physical
resolutions so this properly excludes options that wouldn't be able to fit on the screen once scaled.


Former-commit-id: c3e28d931a4183bf98e5e891c229c2cfd2736d63
Former-commit-id: ebe4b5f97408330686b2dfc321fb848464ff4ecd
Former-commit-id: 43363911cec5575175945c5a98b49e74d3c96fd0
Former-commit-id: a586a714a92a3075e207097527fe34904ec8b4f7 [formerly 9f10069fad6f4d6d843970771422216442295ae5]
Former-commit-id: 4520dd5e3a089d4afd7a18734c3e168b5f2f13cf