Allow widget::find_at() to return a scrollbar container
If the pointer wasn't on top of any widget within a scrollbar container,
when GUI2 queried the widget on which a UI event occurred, the answer was
"no widget" and the whole event was discarded.

That was problematic if the player was trying to scroll. It is desired that
scrolling is possible even if the pointer isn't on a widget (as long as
it's somewhere within the scrollbar container).

Now the query will return "scrollbar container" and dispatch the event to
it. That allows the player to scroll anywhere within a scrollbar container.

Also, now the GUI2 event chain includes the target widget itself, not only
its ancestors.

Fixes #1632.


Former-commit-id: 2bfe30b9e3b46788d94f050746d23c136007c23c
Former-commit-id: efa893c096e8c8408e838889acbe0631c04da246
Former-commit-id: 2d92f87bbfa3a6820f099b34b42515b31392cb19
Former-commit-id: 7a907559cb8b4ac5a9ce5bc06284f4d11ad0576b [formerly 7b248f0601be00d8ae7fac33fb598b0012b7a816]
Former-commit-id: 6fd671705fb27ceb0b622052fc33314b7eab1d34