Attempted fix for GUI2 event firing test failures
Commit 6e0f83d9abca5da8bc4c8e78ae9b4a707403346a [formerly 6fd671705fb27ceb0b622052fc33314b7eab1d34] [formerly 7a907559cb8b4ac5a9ce5bc06284f4d11ad0576b [formerly 7b248f0601be00d8ae7fac33fb598b0012b7a816]] [formerly 2d92f87bbfa3a6820f099b34b42515b31392cb19] [formerly efa893c096e8c8408e838889acbe0631c04da246] [formerly 2bfe30b9e3b46788d94f050746d23c136007c23c] broke the test in two ways. First, I intentionally
changed pre and post events to be passed to the target widget as well, not
only its ancestors. The test verified that the events were *not* passed to
the target widget. Fixed by adjusting the test to verify that the target
widget now does receive the event.

Second, I unintentionally stopped the event dispatcher from receiving pre
and post events. Fixed.


Former-commit-id: 7a5650060cb814e7b790a77ed47cb49795a9483a
Former-commit-id: 4d85dbf7a6ee72115fa79b3da650d31fee947de4
Former-commit-id: 65b107bb9f4587cf56df906b817ba5cfd5d06814
Former-commit-id: 4f9e90e1c2f587e0bb9150ac0b277dfd1e1a47bc [formerly 4f05e891a7302da3bf74b72de3f4628056bfa495]
Former-commit-id: 78b61f79430d8d5d3b5d0e8b26432e29b97187fd