Fixed help topics not working right when multiple factions share an id
For example the factions of Default Era and Age of Heroes have the same id's, which caused the faction help topics to have the same id's as well, leading to a faction link on both era pages to lead to the same faction page. Adding an era prefix to the faction topic id's makes them unique and solves the problem.


Former-commit-id: 65887e3b7ad2d5a0c6f5060f201674bac80c113d
Former-commit-id: 2b312f7f4f63f377fa830154fe81cb01232dba91
Former-commit-id: 4698aa5bd68f8260862011aa91387bd280eb6bcf
Former-commit-id: b21afd5bc329335c12136b67b6f76c44b0d3dd5b [formerly 9d7a8affc392acea4b2e4101b2251fd7ffa503b5]
Former-commit-id: a977e40e49a35408ccb97a9236cbe2fac0380ad1