Split texture caches into linear and NN scaled versions
With textures, you can't change render scale quality once they've been created. To rectify this, I've
made the texture caches a map of caches, sorted by scale quality. This should result in no performance
overhead, since an image is simply added to the appropriate quality cache on creation instead of in
a single cache. If a different version of a texture is needed it will be loaded later.

By default, load_texture will return images using nearest neighbor scaling. GUI2 always fetches images
using linear scaling. This is consistent with the old software rendering method. Plus, we don't have
any way (as of now) to specify the render quality on a per-image basis.


Former-commit-id: 2a9fd5d625e07e41b2bacb5c885e75610a3a275b
Former-commit-id: c496f9c10d896c222643a44e6c18b8e7ada2f66e
Former-commit-id: d053d38246cd9f92e2e0f1493f564698eb6ea0e4
Former-commit-id: f3631ade15a6fd1018c45bd6f371d790addb65db [formerly 1c60d67a64e66474f430449a9ec8a46fbd108011]
Former-commit-id: 4b134e54bbf1989c4d7c4564ca7dcec29ee9ed42