GUI2/Addon List: restrict markup to name and installation status fields
We don't want markup in any of the other fields. I still need to find a slightly better solution for the name
field, though. We don't want users to be able to use markup unnecessairly, but we do need markup to highlight
the names of ready-to-upload addons.

This fixes warnings about unescaped ampersands in the Author field.


Former-commit-id: ffec800bda3051dbf40ca84b8ff9b0accb3d8ff8
Former-commit-id: ef46d0e9ec1c4101b78637df98545c81ab2d0944
Former-commit-id: 4c1bc617030927f345fa681815601b7109a263c1
Former-commit-id: af74ae3e72a8b9ea3566de603e85b4c9671f86b3 [formerly 2bd1dbf9f3c556c6e54d8988d2bc28af658cb2ce]
Former-commit-id: cbd632d424cdbc1aab801ab0ce4f84021ddd86b4