Restored scale_surface_legacy for [message] image scaling and ~SCALE
Fixes issue #1570 (artifacts with alpha gradients), also possibly relevant for parts of the discussion of issue #1254.

The version of scale_surface_legacy that I used is the one without alpha thresholding. It still isn't perfect, as it doesn't quite properly weigh each input pixel according to its alpha. It's possible that scale_surface could simply be fixed to not result in artifacts with alpha gradients, making a separate scaling function like this unnecessary to have, but I wasn't able to find a way to do it.

For pixel art without the kind of alpha gradients it cannot handle, scale_surface does produce a better result, but the difference seems very subtle.

Relevant past commits:
4af0e815
5f967950
2df60619


Former-commit-id: 87c9adc056f99410904e979c3bd4eb1bf15f09ae
Former-commit-id: fe77cf02c73a1bc1d6eb914249ad3c8b1f7f0f82
Former-commit-id: 2d58eb349748c941db6838acc7068d8f99e32cef
Former-commit-id: c80c35948f58902c789c35a0d5c54fe48ea17bd4 [formerly 19480e8f0c47dbfdd42d4b72320ef2e9bce0b6a4]
Former-commit-id: ae7c76a28ad486126370ce37a66ff250587e9536