Another large batch of progress, primarily in game_display
This is primarily focused on getting various areas of the game display drawing, such as footstep images
and the terrain defense indicators.

To facilitate the last, I've refactored draw_text_in_hex to utilize floating labels instead of drawing
to hex directly with surfaces. This probably isn't the most optimal solution, since labels are continuously
created and destroyed, but since their text is usually short and cached in the rendered text cache it
does for now.

Also includes some cleanup of deprecated functions in display, and slight reordering of drawing order of
certain elements.


Former-commit-id: d5f3af6d62b0e6487068dc016fc305d492bf6fce
Former-commit-id: 9d4fb32689db6381a5a5ae3fa6df647cb50d10a2
Former-commit-id: 2e5b9a58c1b9902abaa703d30b36dd942c00407d
Former-commit-id: f09d83cc1088f79146d3554418a7cd063de2708d [formerly 269b7110fb5869968632226c84a00a57aed9d4ee]
Former-commit-id: 3a524071c120bc02c9fb31972ddc6016466aac1b