GUI2/Canvas: optimized cached texture management
Instead of throwing out the cached texture every time the canvas was dirty, I've added a flag so that
only happens if its size has actually changed. There still seem to be some issues in GUI2 where canvas
objects keep getting reset, but I'm not sure why. Will have to look into that.

Anyway, the addition of the SDL_RenderClear call also seems to have fixed the weird graphics bleed-through
I was getting when running the game with the OGL driver.

I don't know if the canvas copy or move ctors are needed for anything. I added them for a test, but I'll
leave them for now in case I need them later.


Former-commit-id: 06ced7fe188299f9754f5cd38d928de719e16cc8
Former-commit-id: 96012d9f011e57d8008b77070c77c6ce61f4c0f8
Former-commit-id: b1dda42440f4ebec8bbaff496ea50e6a667a0c52
Former-commit-id: a893b43fc84f93149c2b78163b6449c8cf4d2caf [formerly f27fdeaa10203f5452df78a9701011e1a6d572d2]
Former-commit-id: 207f90cdfb99dba6c66a83946c329671dbb32536