add [cancel_action] implements #1427
this adds a new tag [cancel_action], it currently only has an effect in
move actions. It can be used in enter_hex/exit_hex events to abort the
current movement. It can also be used in moveto events where is
cancels possible following attacks (if the user issued a move+attack
action).

This also changes the behaviour of moveto and enter/exit_hex event so
that they now no longer cancel the movement automatically, also
[allow_undo] no longer has an effect on whether the move is cancelled,
it now only allows undoing of the move.


Former-commit-id: 30aba652396aa6e637d63d4b2ab979ccc1d973b4
Former-commit-id: 72f53509d7486ff839d3e582a67e836de4c8f5c7
Former-commit-id: 09dde2d2c628e2a241edc3306b7640391d91900d
Former-commit-id: 943ca9dd79bdbe131db2e5d5d56d11010385b7b7 [formerly b7e788596fe0b8a5207830ee0be4f3d27aeef88a]
Former-commit-id: 9aeee9470c80109ef4aae61e64f25d9131ecc46e