Added new minimap drawing method that works wholly with textures
I've left the old getMinimap function alone and added a new function that will eventually replace it
that utilizes texture rendering to a given texture instead of relying on surface blits and scaling.
This has the benefit of, well, reducing surface ops, for one, as well as removing the need to create
a new texture from the minimap surface every time it's updated.


Former-commit-id: 951369b2bfed92beb4a9502e80ae49ce1cf33fce
Former-commit-id: fe608c76d0fc48d6619c062103a6a5907b9b96b8
Former-commit-id: 113649023137e118dc23976ffbe2d85fe250a328
Former-commit-id: fb91cbee59c2f2285241a7b188f2f217d247240c [formerly 8c445d0320d20c60e86deab2d02a7efa8a8e91d0]
Former-commit-id: 57e5cf14754af492f59bd8baaa09cf231959ae35