wam: Remove icons' alt-text
Using the original paths given in WML as alt-text does not serve any
purpose other than breaking the layout when the page hasn't finished
loading and half the icons are overlong text. If anyone really needs
those they should use a data attribute instead. The alt attribute is for
accessibility, not for making programmers happy with themselves.


Former-commit-id: bcaa06d76c1325e8c42274598e077f42c5012706
Former-commit-id: dd946bf7fbbb77cbbd550069ab5e674fb804aa37
Former-commit-id: ea507197b4300af05ad70050d709941c97b6bcbc
Former-commit-id: 7d0ad091a179b898edc2191a771f05d011c62b87 [formerly 5eace0e9fe3e9b90ddd448ab3cdff5caa86c898f]
Former-commit-id: 18890c76748b7df5cf010fe906bec02b01b98954