GUI2/Modal Dialog: don't use pop_back to remove window ptr from open window stack
This seems to have been the cause of some crashes and weird behavior with the new command console.
Likely the problem came from the wrong pointer being removed from the ows, but exactly why a modal
dialog was even being opened I don't know.

Still, this is a safer method overall and it guarantees the correct pointer is always removed.


Former-commit-id: 658d7400b63dc89881fbe855c05c4cb0c4999e49
Former-commit-id: 3ea30a036f84c4b84a254dee050000bf4182d658
Former-commit-id: 816da4cb811c5e1b67f0cdd8e7f0e4179dc33773
Former-commit-id: eafea3654f928d3144d1686ba7fd97cfc89fc868 [formerly b17106c8c3579769e5ee6be0805b2ca29c58344b]
Former-commit-id: 8407f7e718703bd9294a1583b0979d5025fe3046