wmlunits: Escape generated URLs and WML-defined output
Some refactoring and a lot of code changes are involved in this.
Basically, wmlunits has always fully trusted its WML input. This has
several implications:

 * URLs such as the ones for unit detail views tend to contain
   unescaped characters which are forbidden in URLs, such as spaces.
   While browsers generally tolerate this well, the result is still
   invalid HTML. The same applies to URLs used for inline CSS (e.g.
   background-image).

 * Most information read from WML such as unit names, descriptions,
   campaign names, era names, and so on, was blindly trusted and could
   allow an attacker to inject malicious HTML into units.wesnoth.org by
   uploading an add-on that would pass the units.wesnoth.org build
   process.

The resulting code might not be extent of regressions, but hopefully
we'll identify them quickly. This commit contains some additional noise
in the form of style fixes around offending lines as well. Backporting
to 1.12 is, as far as I understand, unnecessary since units.wesnoth.org
uses the master branch version of wmlunits, but it might be worth
evaluating whether to issue a warning for the 1 person in the world who
might be running the 1.12 version of the units.wesnoth.org toolchain on
their own facilities (okay, let's face it, nobody would ever do that).


Former-commit-id: 787ee1f981b931059555b7e82684ff53c5f24cdb
Former-commit-id: be2cf3234791d6a99423edde702cbef4db2b9f74
Former-commit-id: 6921384a6886facdadb2382ef95a798da712588b
Former-commit-id: a279c9798fcc50bd504c1992b76b516d629742df [formerly 7e8b133f36c050137908156cb77aebf12b5f2ec0]
Former-commit-id: 06de6d873b84c5385914df4718a5185b1d07b465