GUI2/Canvas: some refactoring to correctly reimplement the shape caching mechanism
The purpose of the shape caching is essentially to allow multiple draws to consecutively add to
previously drawn content. It essentially adds another condition for clearing the texture prior
to rendering.

I also moved the shape list handling into the draw function itself so it's only executed once
if needed instead multiple times when calling functions like set_variable - not that it actually
would have done anything after the first call, but this allows me to keep the clearing-the-texture
code in one place.


Former-commit-id: 5a76720a07b0c025d5be36ceb7d31373f455c4e9
Former-commit-id: 294555c9b14590b742e9009c5c73f0422a0f7c0f
Former-commit-id: 9b1f3653f9284220dafa65a0cf1922021279d198
Former-commit-id: 1be456e831497047cd8d5704cadc958287f2cba7 [formerly b6fd570fc6c1029c044a162c4ffa8bf6c1b34c30]
Former-commit-id: cc3298862979f739d53d819f1a3604c348596e98