wmlunits: Fix issue where the process name becomes "w" instead of "wmlunits"
The basic string type in Python 3 is equivalent to a wchar_t array,
which means that the second byte of every ASCII string is always zero,
causing the string to appear truncated after the first character when
calling C library functions that expect a single or multibyte encoding
(e.g. UTF-8). Just using a byte array does the trick to get a proper
single-byte ASCII string.


Former-commit-id: 33a736b458793b798de838642ae728dd377fd334
Former-commit-id: 82552ff29da8c64eb120e2e03fd5b41eacae7c76
Former-commit-id: d78b406773f9a20a45b5862ab34d6f16e3d439af
Former-commit-id: 8da44a959839b55136f7261282fdd4f298e47002 [formerly 6ca911fdca632f6ef5c99cf99ac2520c72da8831]
Former-commit-id: a141369ee2177048e6bb99573d34ec45cf31ca36