GUI2/Window Builder: removed build_widget function and used std::make_shared directly
This function's name was rather misleading. It didn't actually build a widget, just created a an
object (usually a builder, as it was utilized in REGISTER_WIDGET) with the config argument as its
ctor parameters.

Lamdas render the need of a separate function to pass as an std::function argument obsolete.


Former-commit-id: b9607d15bfcbea09cb8494ea8aea7114425bfd76
Former-commit-id: f1fc3efa389b7a23d27f76c5ef7522a693dfb994
Former-commit-id: 15de7292031395b8fe1117369c2f6a3b5b8c4d60
Former-commit-id: 0f9d9f8ea725a59c69f9188d20ce14c05c815420 [formerly a94c89dca0e0c629137505a0b5d65062bdb4ffc0]
Former-commit-id: 4ad5fd676bc519892b7d0403c796579c141592be