GUI2/Dispatcher: changed dispatcher signals to take a widget reference as a first parameter
This fixes an issue where functions taking `widget` as a parameter couldn't be bound to a signal function
with the first argument passed via placeholder since `std::bind` requires arguments match exactly when using
the placeholders.

The dispatcher's `this` was cast to `widget` before the signal functions were fired anyway, so this doesn't
really constitute any functional change.


Former-commit-id: dd958d4f0c80fe502daa174d07611d8348d7a8d6
Former-commit-id: 062f6cb5329c24533d9f9c30555303e2d9012705
Former-commit-id: 7fe028b4f578cc11270d7c0f267e631efdeb689c
Former-commit-id: c4855ad4a1ab678a67396a451e6e20b08db78c6d [formerly 93564a670fdd2dc0c77fef7d1fe0c445658dc84a]
Former-commit-id: b1fdcd9e6a93fdf34476a7a066cdb56c2e545616