Preprocessor: refactor streambuf management of preprocessor objects
Instead of using this complex system where new processor objects registered themselves as the
current preprocessor object in the base class ctor (this led to freestanding `new` statements
that looked like memory leaks), we now use a deque of unique_ptrs in the streambuf class. This
makes the code a lot clearer since all the memory management is is one place.


Former-commit-id: 9e16e192af483483ce41f71ae05552875d62245f
Former-commit-id: 0e9775e978d77fc308f11b2519af9660d5ccd7a4
Former-commit-id: 2172ff1b4f19236f3357595a28ddc96ebed1d28d
Former-commit-id: ef0369dc0e909aa9c1c7c3d17a78a4baa5461b4f [formerly 4b2ca1b9e4c0b580f22ce34e1fe47319111e6c7c]
Former-commit-id: 34aa972d25a177d1c0f1d7482be92fe837219324